CXBOX-907 | fix interfaces and Dictionary
