+++ ui/src/components/Card/Card.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Col, Row } from 'antd'
 import cn from 'classnames'
-import { interfaces } from '@cxbox-ui/core'
+import { WidgetTypes } from '@cxbox-ui/core'
 import Operations from '../Operations/Operations'
 import { useAppSelector } from '@store'
 import { useWidgetCollapse } from '@hooks/useWidgetCollapse'
@@ -16,8 +16,6 @@ export interface CardProps {
     className?: string
 }
 
-const { WidgetTypes } = interfaces
-
 const showOperations = [WidgetTypes.DataGrid, WidgetTypes.Form]
 
 function Card({ meta, children, className }: CardProps) {
@@ -26,7 +24,6 @@ function Card({ meta, children, className }: CardProps) {
 
     const bcUrl = useAppSelector(state => state.screen.bo.bc[bcName] && buildBcUrl(bcName, true))
     const operations = useAppSelector(state => state.view.rowMeta?.[bcName]?.[bcUrl]?.actions)
-    const debugMode = useAppSelector(state => state.session.debugMode || false)
     const isForm = type === WidgetTypes.Form
 
     return (
@@ -44,7 +41,7 @@ function Card({ meta, children, className }: CardProps) {
                     {!(isMainWidget && isCollapsed) && (
                         <>
                             {isForm && children}
-                            {showOperations.includes(type as interfaces.WidgetTypes) && (
+                            {showOperations.includes(type as WidgetTypes) && (
                                 <Operations
                                     operations={operations}
                                     bcName={bcName}
