+++ ui/src/components/Popup/Popup.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next'
 import styles from './Popup.less'
 import { ModalProps } from 'antd/lib/modal'
 import cn from 'classnames'
-import Pagination from '../ui/Pagination/Pagination'
-import Button from '../ui/Button/Button'
+import Button from '@components/ui/Button/Button'
+import Pagination from '@components/ui/Pagination/Pagination'
 import { interfaces } from '@cxbox-ui/core'
 import { useAppSelector } from '@store'
